Move zarrs_to_dataset to sgkit.io.utils to avoid implicit cyvcf2 dependency by vcfzarr_to_zarr
