+++ sgkit/io/vcf/__init__.py
@@ -1,8 +1,9 @@
 import platform
 
 try:
+    from ..utils import zarrs_to_dataset
     from .vcf_partition import partition_into_regions
-    from .vcf_reader import vcf_to_zarr, vcf_to_zarrs, zarrs_to_dataset
+    from .vcf_reader import vcf_to_zarr, vcf_to_zarrs
 
     __all__ = [
         "partition_into_regions",
